# app/routes/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AnimaLink - OpenAPI 3.0",
    "description": "Documentation of Backend Apis for AnimaLink Corp",
    "license": {
      "name": "MIT License",
      "url": "https://github.com/AnimaLink/backend-api/blob/main/LICENSE"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/forum/types": {
      "post": {
        "tags": ["forum-aditionals"],
        "summary": "Create new forum type",
        "description": "In case you dont want to send files in <b>attachment</b> field, please uncheck <b>Send empty value</b>",
        "operationId": "createForumType",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "attachment": {
                    "type": "string",
                    "format": "binary"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "create forum type successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "create forum type success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "forumStatusId": {
                          "type": "integer",
                          "example": "1"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "array",
                          "example": [
                            "email must be an email",
                            "password must be minimum 6 characters"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "object",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Something went wrong please try again later."
                    },
                    "data": {
                      "type": "object",
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["forum-aditionals"],
        "summary": "Get all forum status",
        "description": "",
        "operationId": "getAllForumStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "get all forum status success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "get all forum status success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "listForumStatus": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "example": 1
                              },
                              "name": {
                                "type": "string",
                                "example": "Available"
                              },
                              "imgUrl": {
                                "type": "string",
                                "example": "https://storage.googleapis.com/AWOWKOWKOWK/20231116-162039"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "data": {
                      "type": "object",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Something went wrong please try again later."
                    },
                    "data": {
                      "type": "object",
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}